requirements.txt
langchain
langchain-google-genai
langgraph
playwright
python-dotenv